+++ src/Client/Driver/DoctrineCacheDriverDecorator.php
@@ -3,6 +3,7 @@
 namespace Gos\Bundle\WebSocketBundle\Client\Driver;
 
 use Doctrine\Common\Cache\Cache;
+use Doctrine\Common\Cache\ClearableCache;
 use Symfony\Component\Cache\DoctrineProvider;
 
 trigger_deprecation('gos/web-socket-bundle', '3.4', 'The "%s" class is deprecated and will be removed in 4.0, use the "%s" class with a "%s" instance instead.', DoctrineCacheDriverDecorator::class, SymfonyCacheDriverDecorator::class, DoctrineProvider::class);
@@ -45,5 +46,12 @@ final class DoctrineCacheDriverDecorator implements DriverInterface
     public function delete(string $id): bool
     {
         return $this->cacheProvider->delete($id);
+    }
+
+    public function clear(): void
+    {
+        if ($this->cacheProvider instanceof ClearableCache) {
+            $this->cacheProvider->deleteAll();
+        }
     }
 }